fix: width값 Integer로 수정 및 기타 수정

+++ Backend/src/main/java/com/potless/backend/project/dto/response/DamageDetailToProjectDto.java
@@ -27,7 +27,7 @@ public class DamageDetailToProjectDto {
     @Schema(description = "주소")
     private String address;
     @Schema(description = "너비")
-    private Double width;
+    private Integer width;
     @Schema(description = "작업 상태")
     private Status status;
     @Schema(description = "지역 구")
@@ -51,7 +51,7 @@ public class DamageDetailToProjectDto {
         this.dirX = dirX;
         this.dirY = dirY;
         this.address = address;
-        this.width = width;
+        this.width = width.intValue();
         this.status = status;
         this.area = area;
         this.location = location;
@@ -68,7 +68,7 @@ public class DamageDetailToProjectDto {
         this.dirX = dirX;
         this.dirY = dirY;
         this.address = address;
-        this.width = width;
+        this.width = width.intValue();
         this.status = status;
         this.area = area;
         this.location = location;
